US-103-703 - Perfil del subscriptor y filtrar ejercicios por tipo

+++ frontend/src/app/views/pages/forms/measurement/edit/editMeasurement.component.ts
@@ -2,6 +2,7 @@ import {Component, OnInit} from '@angular/core';
 import {ActivatedRoute, Router} from '@angular/router';
 
 import {MeasurementService} from '../../../../../services/measurement.service';
+import Swal from "sweetalert2";
 
 @Component({
   selector: 'app-edit-measurement-component',
@@ -12,11 +13,13 @@ export class EditMeasurementComponent implements OnInit {
   subscribers: any;
   currentMeasurement = null;
   message = '';
+
   constructor(
     private measurementService: MeasurementService,
     private route: ActivatedRoute,
     private router: Router) {
   }
+
   ngOnInit(): void {
     this.message = '';
     this.getMeasurement(this.route.snapshot.paramMap.get('id'));
@@ -39,7 +42,15 @@ export class EditMeasurementComponent implements OnInit {
       .subscribe(
         response => {
           console.log(response);
-          this.message = 'Las medidas fueron actualizadas';
+          Swal.fire(
+            {
+              position: 'center',
+              showConfirmButton: false,
+              timer: 1500,
+              icon: 'success',
+              title: 'Las medidas fueron actualizadas'
+            }
+          )
         },
         error => {
           console.log(error);
@@ -47,15 +58,28 @@ export class EditMeasurementComponent implements OnInit {
   }
 
   deleteMeasurement(): void {
-    const subId = this.currentMeasurement.subscriber.id;
-    this.measurementService.delete(this.currentMeasurement.id)
-      .subscribe(
-        response => {
-          console.log(response);
-          this.router.navigate(['/subscriber-profile/'+subId]);
-        },
-        error => {
-          console.log(error);
-        });
+    Swal.fire(
+      {
+        showCancelButton: true,
+        confirmButtonText: 'Eliminar',
+        cancelButtonText: 'Cancelar',
+        icon: 'warning',
+        title: '¿Desea eliminar las medidas?'
+      }
+    ).then(result => {
+      const subId = this.currentMeasurement.subscriber.id;
+      this.measurementService.delete(this.currentMeasurement.id)
+        .subscribe(
+          response => {
+            console.log(response);
+            Swal.fire('Eliminado', 'Se eliminaron las medidas.', 'info')
+              .then(result => {
+                this.router.navigate(['/subscriber-profile/' + subId])
+              });
+          },
+          error => {
+            console.log(error);
+          });
+    })
   }
 }
